LayoutTests/fast/events/ios/show-keyboard-when-becoming-first-responder-despite-inputmode-none.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script src="../../../resources/ui-helper.js"></script>
<script src="resources/check-keyboard-on-screen.js"></script>
</head>
<body>
<input id="input">
<script>
window.jsTestIsAsync = true;

let inputElement = document.getElementById("input");

async function runTest()
{
    if (!window.testRunner) {
        testFailed("Must be run in WebKitTestRunner.");
        return;
    }

    await UIHelper.setHardwareKeyboardAttached(false);

    debug("<br>After tapping &lt;input&gt;:");
    await UIHelper.activateElementAndWaitForInputSession(inputElement);
    await checkKeyboardOnScreen();

    debug("<br>After resigning first responder:");
    await UIHelper.resignFirstResponder();
    await UIHelper.waitForKeyboardToHide();
    await checkKeyboardNotOnScreen();

    debug("<br>After becoming first responder:");
    await UIHelper.becomeFirstResponder();
    await UIHelper.waitForKeyboardToShow();
    await checkKeyboardOnScreen();

    document.body.removeChild(inputElement);
    finishJSTest();
}

inputElement.addEventListener("focus", () => inputElement.setAttribute("inputmode", "none"), true /* capture phase */);

description("This tests that in a focused text field the software keyboard is hidden and shown when the web view resigns and becomes first responder, respectively, eventhough focusing the text field sets inputmode to &quot;none&quot;.");
runTest();
</script>
</body>
</html>
